Confiner + state driven camera
Added a confiner which is the boundary limit for where the camera can go (similar to clipping against walls and preventing you from seeing past a point. (we don't really have a tilemap setup so this will need to be setup later.
there is a confiner component added to the camera, when the tilemap basics are ready we can setup the confiner properly. right now its just an empty value.

Note we will need to go into the project settings and in physics 2D edit the matrix so that the background can't collide with anything

State driven camera  - Nothing really setup here but if we want the camera to edit based off certain actions (cutscenes, dashing effect etc).
Created some generic states
By adding the statedriven i have moved the follow camera into state driven

Created some folders for basic organization

Prefabbed the main player and added an animator component in order to properly setup camera.

Setup some state transitions you will see on the state driven camera there are state instructions, This goes in hand with the animator and animation transitions I added.
TLDR as you switch to an animation the camera will adjust accordingly so you can get smooth animation transitions and a really nice feel for whatever movement or cinematic action is going on .